Move `IntervalPrinter` to a more suitable location.
